Add elemental effects to Bullets and health system
Introduces fire and ice effects to projectiles, allowing them to apply damage-over-time (DoT) or freeze targets. Health system now supports float values, DoT, and freeze states, with corresponding changes to interfaces and UI. Enemy and patrol logic updated to respect freeze state. Prefabs and scene updated to configure new projectile properties.

+++ Assets/Scripts/Gameplay/Combat/IDamageable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void ApplyDamage(int amount, GameObject source);
+    void ApplyDamage(float amount, GameObject source);
 }
